Extract CardEventPublisher class as a collaborator with MatrixService
