+++ packages/host/app/services/matrix-service.ts
@@ -1,4 +1,4 @@
-import type Owner from '@ember/owner';
+import { getOwner } from '@ember/owner';
 import type RouterService from '@ember/routing/router-service';
 import { debounce } from '@ember/runloop';
 import Service, { service } from '@ember/service';
@@ -12,24 +12,18 @@ import {
   type MatrixEvent,
   type RoomMember,
   type EmittedEvents,
-  type ISendEventResponse,
 } from 'matrix-js-sdk';
 import stringify from 'safe-stable-stringify';
-import { md5 } from 'super-fast-md5';
 import { TrackedMap } from 'tracked-built-ins';
 import { v4 as uuidv4 } from 'uuid';
 
 import {
   aiBotUsername,
   baseRealm,
-  codeRefWithAbsoluteURL,
-  getClass,
   LooseCardResource,
   logger,
   markdownToHtml,
   ResolvedCodeRef,
-  splitStringIntoChunks,
-  type LooseSingleCardDocument,
 } from '@cardstack/runtime-common';
 
 import {
@@ -42,9 +36,7 @@ import { escapeHtmlOutsideCodeBlocks } from '@cardstack/runtime-common/helpers/h
 import { getMatrixUsername } from '@cardstack/runtime-common/matrix-client';
 
 import {
-  APP_BOXEL_CARD_FORMAT,
   APP_BOXEL_CARDFRAGMENT_MSGTYPE,
-  APP_BOXEL_COMMAND_DEFINITIONS_MSGTYPE,
   APP_BOXEL_COMMAND_RESULT_EVENT_TYPE,
   APP_BOXEL_COMMAND_RESULT_REL_TYPE,
   APP_BOXEL_COMMAND_RESULT_WITH_NO_OUTPUT_MSGTYPE,
@@ -56,7 +48,6 @@ import {
   APP_BOXEL_ACTIVE_LLM,
   DEFAULT_LLM_LIST,
   APP_BOXEL_COMMAND_REQUESTS_KEY,
-  APP_BOXEL_ROOM_SKILLS_EVENT_TYPE,
 } from '@cardstack/runtime-common/matrix-constants';
 
 import {
@@ -69,12 +60,7 @@ import Room, { TempEvent } from '@cardstack/host/lib/matrix-classes/room';
 import { getRandomBackgroundURL, iconURLFor } from '@cardstack/host/lib/utils';
 import { getMatrixProfile } from '@cardstack/host/resources/matrix-profile';
 
-import type { Base64ImageField as Base64ImageFieldType } from 'https://cardstack.com/base/base64-image';
-import type {
-  relativeTo,
-  BaseDef,
-  CardDef,
-} from 'https://cardstack.com/base/card-api';
+import type { BaseDef, CardDef } from 'https://cardstack.com/base/card-api';
 import type * as CardAPI from 'https://cardstack.com/base/card-api';
 import type * as FileAPI from 'https://cardstack.com/base/file-api';
 import { type FileDef } from 'https://cardstack.com/base/file-api';
@@ -86,13 +72,13 @@ import type {
   CommandResultWithOutputContent,
   RealmEventContent,
   CommandDefinitionsContent,
-  CommandDefinitionSchema,
+  Tool,
 } from 'https://cardstack.com/base/matrix-event';
 
-import type { Tool } from 'https://cardstack.com/base/matrix-event';
 import type * as SkillCardModule from 'https://cardstack.com/base/skill-card';
 
 import AddSkillsToRoomCommand from '../commands/add-skills-to-room';
+import CardEventPublisher from '../lib/card-event-publisher';
 import { importResource } from '../resources/import';
 
 import { RoomResource, getRoom } from '../resources/room';
@@ -111,12 +97,10 @@ import type NetworkService from './network';
 import type RealmService from './realm';
 import type RealmServerService from './realm-server';
 import type ResetService from './reset';
-import type { Skill as RoomSkill } from '../components/ai-assistant/skill-menu';
 
 import type * as MatrixSDK from 'matrix-js-sdk';
 
 const { matrixURL } = ENV;
-const MAX_CARD_SIZE_KB = 60;
 const STATE_EVENTS_OF_INTEREST = ['m.room.create', 'm.room.name'];
 
 const realmEventsLogger = logger('realm:events');
@@ -147,6 +131,10 @@ export default class MatrixService extends Service {
 
   private roomDataMap: TrackedMap<string, Room> = new TrackedMap();
   private startedAtTs = -1;
+  private cardEventPublisher = new CardEventPublisher(
+    getOwner(this),
+    () => this.cardAPI,
+  );
 
   roomResourcesCache: TrackedMap<string, RoomResource> = new TrackedMap();
   messagesToSend: TrackedMap<string, string | undefined> = new TrackedMap();
@@ -166,8 +154,6 @@ export default class MatrixService extends Service {
   #eventBindings: [EmittedEvents, (...arg: any[]) => void][] | undefined;
   currentUserEventReadReceipts: TrackedMap<string, { readAt: Date }> =
     new TrackedMap();
-  private cardHashes: Map<string, string> = new Map(); // hashes <> event id
-  private commandDefHashes: string[] = []; // hashes
 
   constructor(owner: Owner) {
     super(owner);
@@ -562,6 +548,25 @@ export default class MatrixService extends Service {
     });
   }
 
+  async addCardsToRoom(cards: CardDef[], roomId: string) {
+    let cardEventIds = await this.cardEventPublisher.addCardsToRoom(
+      cards,
+      roomId,
+    );
+    return cardEventIds;
+  }
+
+  async addSkillCardsToRoomHistory(
+    skillCards: SkillCardModule.SkillCard[],
+    roomId: string,
+  ) {
+    let cardEventIds = await this.cardEventPublisher.addSkillCardsToRoomHistory(
+      skillCards,
+      roomId,
+    );
+    return cardEventIds;
+  }
+
   async sendCommandResultEvent(
     roomId: string,
     invokedToolFromEventId: string,
@@ -570,7 +575,10 @@ export default class MatrixService extends Service {
   ) {
     let resultCardEventId: string | undefined;
     if (resultCard) {
-      [resultCardEventId] = await this.addCardsToRoom([resultCard], roomId);
+      [resultCardEventId] = await this.cardEventPublisher.addCardsToRoom(
+        [resultCard],
+        roomId,
+      );
     }
     let content:
       | CommandResultWithNoOutputContent
@@ -612,185 +620,6 @@ export default class MatrixService extends Service {
         }`,
       );
     }
-  }
-
-  async addCommandDefinitionsToRoomHistory(
-    commandDefinitions: SkillCardModule.CommandField[],
-    roomId: string,
-  ) {
-    console.trace(
-      'addCommandDefinitionsToRoomHistory',
-      commandDefinitions,
-      roomId,
-    );
-    // Create the command defs so getting the json schema
-    // and send it to the matrix room.
-    let commandDefinitionSchemas: CommandDefinitionSchema[] = [];
-    const mappings = await basicMappings(this.loaderService.loader);
-    for (let commandDef of commandDefinitions) {
-      let absoluteCodeRef = codeRefWithAbsoluteURL(
-        commandDef.codeRef,
-        commandDef[Symbol.for('cardstack-relative-to') as typeof relativeTo],
-      ) as ResolvedCodeRef;
-      const Command = await getClass(
-        absoluteCodeRef,
-        this.loaderService.loader,
-      );
-      const command = new Command(this.commandService.commandContext);
-      const name = commandDef.functionName;
-      const schema: CommandDefinitionSchema = {
-        codeRef: absoluteCodeRef,
-        tool: {
-          type: 'function' as Tool['type'],
-          function: {
-            name,
-            description: command.description,
-            parameters: {
-              type: 'object',
-              properties: {
-                description: {
-                  type: 'string',
-                },
-                ...(await command.getInputJsonSchema(this.cardAPI, mappings)),
-              },
-              required: ['attributes', 'description'],
-            },
-          },
-        },
-      };
-      let hashKey = this.generateCommandDefHashKey(roomId, schema);
-      if (!this.commandDefHashes.includes(hashKey)) {
-        commandDefinitionSchemas.push(schema);
-        this.commandDefHashes.push(hashKey);
-      }
-    }
-    if (commandDefinitionSchemas.length) {
-      await this.sendEvent(roomId, 'm.room.message', {
-        msgtype: APP_BOXEL_COMMAND_DEFINITIONS_MSGTYPE,
-        body: 'Command Definitions',
-        data: {
-          commandDefinitions: commandDefinitionSchemas,
-        },
-      });
-    }
-  }
-
-  async addSkillCardsToRoomHistory(
-    skills: SkillCardModule.SkillCard[],
-    roomId: string,
-  ): Promise<string[]> {
-    const commandDefinitions = skills.flatMap((skill) => skill.commands);
-    if (commandDefinitions.length) {
-      await this.addCommandDefinitionsToRoomHistory(commandDefinitions, roomId);
-    }
-    return this.addCardsToRoom(skills, roomId);
-  }
-
-  async addCardsToRoom(cards: CardDef[], roomId: string): Promise<string[]> {
-    if (!cards.length) {
-      return [];
-    }
-    let cardEntries: {
-      card: CardDef;
-      serialization: LooseSingleCardDocument;
-      eventId?: string;
-      wasPreviouslySaved?: boolean;
-      matchingRoomSkill?: RoomSkill;
-    }[] = [];
-    cardEntries = await Promise.all(
-      cards.map(async (card) => {
-        let opts: CardAPI.SerializeOpts = { useAbsoluteURL: true };
-        if (isSkillCard(card)) {
-          opts['includeComputeds'] = true;
-        }
-
-        let { Base64ImageField } = await this.loaderService.loader.import<{
-          Base64ImageField: typeof Base64ImageFieldType;
-        }>(`${baseRealm.url}base64-image`);
-        let serialization = await this.cardService.serializeCard(card, {
-          omitFields: [Base64ImageField],
-          ...opts,
-        });
-        return { card, serialization };
-      }),
-    );
-
-    if (cardEntries.length) {
-      for (let entry of cardEntries) {
-        console.log(JSON.stringify(entry.serialization, null, 2));
-        let hashKey = this.generateCardHashKey(roomId, entry.serialization);
-        let eventId = this.cardHashes.get(hashKey);
-        if (eventId) {
-          entry.wasPreviouslySaved = true;
-        } else {
-          entry.wasPreviouslySaved = false;
-          let responses = await this.sendCardFragments(
-            roomId,
-            entry.serialization,
-          );
-          eventId = responses[0].event_id; // we only care about the first fragment
-          this.cardHashes.set(hashKey, eventId!);
-        }
-        entry.eventId = eventId;
-      }
-    }
-    const skillCardEntries = cardEntries.filter((entry) =>
-      isSkillCard(entry.card),
-    );
-    const roomResource = this.roomResourcesCache.get(roomId);
-    const roomSkills = roomResource?.skills ?? [];
-    for (const skillCardEntry of skillCardEntries) {
-      skillCardEntry.matchingRoomSkill = roomSkills.find(
-        (roomSkill) => roomSkill.card.id === skillCardEntry.card.id,
-      );
-      if (skillCardEntry.matchingRoomSkill) {
-        let commandDefinitions = (
-          skillCardEntry.card as SkillCardModule.SkillCard
-        ).commands;
-        if (commandDefinitions.length) {
-          await this.addCommandDefinitionsToRoomHistory(
-            commandDefinitions,
-            roomId,
-          );
-        }
-      }
-    }
-    let savedSkillCardEntries = skillCardEntries.filter(
-      (entry) => !entry.wasPreviouslySaved,
-    );
-    if (savedSkillCardEntries.some((entry) => entry.matchingRoomSkill)) {
-      await this.updateStateEvent(
-        roomId,
-        APP_BOXEL_ROOM_SKILLS_EVENT_TYPE,
-        '',
-        async (currentSkillsConfig) => {
-          let newSkillsConfig = {
-            enabledEventIds: [...(currentSkillsConfig.enabledEventIds || [])],
-            disabledEventIds: [...(currentSkillsConfig.disabledEventIds || [])],
-          };
-          for (const skillCardEntry of savedSkillCardEntries) {
-            if (skillCardEntry.matchingRoomSkill) {
-              // replace the old skillEventId with the new one
-              newSkillsConfig.enabledEventIds =
-                newSkillsConfig.enabledEventIds.map((eventId: string) =>
-                  eventId === skillCardEntry.matchingRoomSkill!.skillEventId
-                    ? skillCardEntry.eventId!
-                    : eventId,
-                );
-              newSkillsConfig.disabledEventIds =
-                newSkillsConfig.disabledEventIds.map((eventId: string) =>
-                  eventId === skillCardEntry.matchingRoomSkill!.skillEventId
-                    ? skillCardEntry.eventId!
-                    : eventId,
-                );
-            }
-          }
-          return newSkillsConfig;
-        },
-      );
-    }
-
-    return cardEntries.map((entry) => entry.eventId!);
   }
 
   async uploadFiles(files: FileDef[]) {
@@ -859,7 +688,7 @@ export default class MatrixService extends Service {
       }
     }
 
-    let attachedCardsEventIds = await this.addCardsToRoom(
+    let attachedCardsEventIds = await this.cardEventPublisher.addCardsToRoom(
       attachedCards,
       roomId,
     );
@@ -880,47 +709,6 @@ export default class MatrixService extends Service {
         },
       },
     } as CardMessageContent);
-  }
-
-  private generateCardHashKey(roomId: string, card: LooseSingleCardDocument) {
-    return md5(roomId + JSON.stringify(card));
-  }
-
-  private generateCommandDefHashKey(
-    roomId: string,
-    commandDefSchema: CommandDefinitionsContent['data']['commandDefinitions'][0],
-  ) {
-    return md5(roomId + JSON.stringify(commandDefSchema));
-  }
-
-  private async sendCardFragments(
-    roomId: string,
-    card: LooseSingleCardDocument,
-  ): Promise<ISendEventResponse[]> {
-    let fragments = splitStringIntoChunks(
-      JSON.stringify(card),
-      MAX_CARD_SIZE_KB,
-    );
-    let responses: ISendEventResponse[] = [];
-    for (let index = fragments.length - 1; index >= 0; index--) {
-      let cardFragment = fragments[index];
-      let response = await this.sendEvent(roomId, 'm.room.message', {
-        msgtype: APP_BOXEL_CARDFRAGMENT_MSGTYPE,
-        format: APP_BOXEL_CARD_FORMAT,
-        body: `card fragment ${index + 1} of ${fragments.length}`,
-        formatted_body: `card fragment ${index + 1} of ${fragments.length}`,
-        data: {
-          ...(index < fragments.length - 1
-            ? { nextFragment: responses[0].event_id }
-            : {}),
-          cardFragment,
-          index,
-          totalParts: fragments.length,
-        },
-      } as CardFragmentContent);
-      responses.unshift(response);
-    }
-    return responses;
   }
 
   getLastActiveTimestamp(roomId: string, defaultTimestamp: number) {
@@ -1043,7 +831,10 @@ export default class MatrixService extends Service {
     this.flushRoomState = undefined;
     this.unbindEventListeners();
     this._client = this.matrixSDK.createClient({ baseUrl: matrixURL });
-    this.cardHashes = new Map();
+    this.cardEventPublisher = new CardEventPublisher(
+      getOwner(this),
+      () => this.cardAPI,
+    );
     this._currentRoomId = undefined;
     this.messagesToSend = new TrackedMap();
     this.cardsToSend = new TrackedMap();
@@ -1655,24 +1446,4 @@ function getAuth(): LoginResponse | undefined {
     return;
   }
   return JSON.parse(auth) as LoginResponse;
-}
-
-// we check this way because instanceof is unreliable with different module contexts
-// TODO: consider a static symbol instead!
-function isSkillCard(card: CardDef): boolean {
-  // Start with the current card's constructor
-  let proto = Object.getPrototypeOf(card);
-
-  // Walk up the prototype chain
-  while (proto) {
-    const constructor = proto.constructor;
-    // Check if the constructor name is "SkillCard"
-    if (constructor && constructor.name === 'SkillCard') {
-      return true;
-    }
-    // Move up the prototype chain
-    proto = Object.getPrototypeOf(proto);
-  }
-
-  return false;
 }
